+++ jcloud-uaa/src/main/java/com/jwk/uaa/constant/JwkOIDCConstant.java
@@ -1,0 +1,47 @@
+package com.jwk.uaa.constant;
+
+/**
+ * @author Jiwk
+ * @version 0.1.4
+ * <p>
+ * OIDC常量
+ * @date 2022/12/1
+ */
+public interface JwkOIDCConstant {
+
+  /**
+   * 图标
+   */
+  String ICON = "icon";
+
+  /**
+   * 用户名
+   */
+  String USERNAME = "username";
+
+  /**
+   * 手机号
+   */
+  String PHONE = "phone";
+
+  /**
+   * 加密类型
+   */
+  String KEY_STORE_TYPE = "jks";
+
+  /**
+   * 别名
+   */
+  String KEY_STORE_ALIAS = "jose";
+
+  /**
+   * 加密密码
+   */
+  String KEY_STORE_STORE_PASS = "jcloud.cn";
+
+  /**
+   * 加密证书地址
+   */
+  String KEY_STORE_PATH = "static/jose/jose.jks";
+
+}