+++ jwk-common/jwk-common-security/src/main/java/com/jwk/common/security/support/service/JwkRedisOAuth2AuthorizationService.java
@@ -1,23 +1,22 @@
 package com.jwk.common.security.support.service;
 
+import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.RedisSerializer;
 import org.springframework.lang.Nullable;
 import org.springframework.security.oauth2.core.OAuth2AccessToken;
 import org.springframework.security.oauth2.core.OAuth2AuthorizationCode;
 import org.springframework.security.oauth2.core.OAuth2RefreshToken;
 import org.springframework.security.oauth2.core.OAuth2TokenType;
 import org.springframework.security.oauth2.core.endpoint.OAuth2ParameterNames;
+import org.springframework.security.oauth2.core.oidc.OidcIdToken;
 import org.springframework.security.oauth2.server.authorization.OAuth2Authorization;
 import org.springframework.security.oauth2.server.authorization.OAuth2AuthorizationService;
 import org.springframework.util.Assert;
-
-import java.time.temporal.ChronoUnit;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.concurrent.TimeUnit;
 
 /**
  * @author Jiwk
@@ -57,16 +56,30 @@ public class JwkRedisOAuth2AuthorizationService implements OAuth2AuthorizationSe
 
 		if (isRefreshToken(authorization)) {
 			OAuth2RefreshToken refreshToken = authorization.getRefreshToken().getToken();
-			long between = ChronoUnit.SECONDS.between(refreshToken.getIssuedAt(), refreshToken.getExpiresAt());
-			redisTemplate.opsForValue().set(buildKey(OAuth2ParameterNames.REFRESH_TOKEN, refreshToken.getTokenValue()),
-					authorization, between, TimeUnit.SECONDS);
+			if (refreshToken.getExpiresAt() != null) {
+				long between = ChronoUnit.SECONDS.between(refreshToken.getIssuedAt(), refreshToken.getExpiresAt());
+				redisTemplate.opsForValue().set(buildKey(OAuth2ParameterNames.REFRESH_TOKEN, refreshToken.getTokenValue()),
+								authorization, between, TimeUnit.SECONDS);
+			}
 		}
 
 		if (isAccessToken(authorization)) {
-			OAuth2AccessToken accessToken = authorization.getAccessToken().getToken();
-			long between = ChronoUnit.SECONDS.between(accessToken.getIssuedAt(), accessToken.getExpiresAt());
-			redisTemplate.opsForValue().set(buildKey(OAuth2ParameterNames.ACCESS_TOKEN, accessToken.getTokenValue()),
-					authorization, between, TimeUnit.SECONDS);
+				OAuth2AccessToken accessToken = authorization.getAccessToken().getToken();
+				if (accessToken.getExpiresAt() != null) {
+				long between = ChronoUnit.SECONDS.between(accessToken.getIssuedAt(), accessToken.getExpiresAt());
+				redisTemplate.opsForValue().set(buildKey(OAuth2ParameterNames.ACCESS_TOKEN, accessToken.getTokenValue()),
+								authorization, between, TimeUnit.SECONDS);
+			}
+		}
+
+		if (isIdToken(authorization)) {
+			OAuth2Authorization.Token<OidcIdToken> oidcIdTokenAuthorization = authorization.getToken(OidcIdToken.class);
+			OidcIdToken oidcIdToken = oidcIdTokenAuthorization.getToken();
+				if (oidcIdToken.getExpiresAt() != null) {
+				long between = ChronoUnit.SECONDS.between(oidcIdToken.getIssuedAt(), oidcIdToken.getExpiresAt());
+				redisTemplate.opsForValue().set(buildKey(OAuth2ParameterNames.ACCESS_TOKEN, oidcIdToken.getTokenValue()),
+								authorization, between, TimeUnit.SECONDS);
+			}
 		}
 	}
 
@@ -95,6 +108,13 @@ public class JwkRedisOAuth2AuthorizationService implements OAuth2AuthorizationSe
 		if (isAccessToken(authorization)) {
 			OAuth2AccessToken accessToken = authorization.getAccessToken().getToken();
 			keys.add(buildKey(OAuth2ParameterNames.ACCESS_TOKEN, accessToken.getTokenValue()));
+		}
+
+		if (isIdToken(authorization)) {
+			OAuth2Authorization.Token<OidcIdToken> oidcIdTokenAuthorization = authorization
+					.getToken(OidcIdToken.class);
+			OidcIdToken oidcIdToken = oidcIdTokenAuthorization.getToken();
+			keys.add(buildKey(OAuth2ParameterNames.ACCESS_TOKEN, oidcIdToken.getTokenValue()));
 		}
 		redisTemplate.delete(keys);
 	}
@@ -133,6 +153,12 @@ public class JwkRedisOAuth2AuthorizationService implements OAuth2AuthorizationSe
 
 	private static boolean isAccessToken(OAuth2Authorization authorization) {
 		return Objects.nonNull(authorization.getAccessToken());
+	}
+
+	private static boolean isIdToken(OAuth2Authorization authorization) {
+		OAuth2Authorization.Token<OidcIdToken> authorizationCode = authorization
+				.getToken(OidcIdToken.class);
+		return Objects.nonNull(authorizationCode);
 	}
 
 }